Split crc in a separate file and added entries for missing SBAS messages.
